requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
pydantic==2.11.4
pydantic-settings==2.9.1
python-dotenv==1.1.0
email_validator==2.2.0
alembic==1.16.1
asyncpg==0.30.0
SQLAlchemy==2.0.41
psycopg2-binary==2.9.10
psycopg2==2.9.10
